clang-tidy integration. Not using it in CI yet because there are certain unresolved failures that will be taken care of when 64-bit platforms are supported.

+++ canard_internals.h
@@ -101,7 +101,7 @@ CANARD_INTERNAL int16_t descatterTransferPayload(const CanardRxTransfer* transfe
 
 CANARD_INTERNAL bool isBigEndian(void);
 
-CANARD_INTERNAL void swapByteOrder(void* data, unsigned size);
+CANARD_INTERNAL void swapByteOrder(void* data, size_t size);
 
 CANARD_INTERNAL uint16_t crcAddByte(uint16_t crc_val, uint8_t byte);
 CANARD_INTERNAL uint16_t crcAdd(uint16_t crc_val, const uint8_t* bytes, size_t len);
